refactor : 사용하지 않는 dag 제거
